Agent_Templates/Runtime_env/pyproject.toml
[tool.poetry]
name = "agentsmithy-starter-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <your@email.com>"]
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "0.110.3"
pydantic = "^2.10.0"
toml = "^0.10.2"
PyYAML = "^6.0.2"
cloudpickle = "3.0.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
langchain-google-vertexai = "^2.0.7"
langchain = "^0.3.0"
langgraph = "^0.2.69"
langchain-core = "^0.3.9"
langchain-community = "^0.3.4"
langchain-google-community = {extras = ["vertexaisearch"], version = "^2.0.2"}
traceloop-sdk = "^0.33.12"
opentelemetry-exporter-gcp-trace = "^1.6.0"
opentelemetry-sdk = "^1.28.0"
google-cloud-logging = "^3.10.0"
google-cloud-aiplatform = {extras = ["evaluation", "agent_engines", "langchain"], version = "^1.84.0"}
google-cloud-discoveryengine = "^0.11.14"
grpcio = "1.70.0"
scikit-learn = "^1.5.0"
immutabledict = "^4.2.0"
pypdf = "^4.3.1"
langchain-openai = "^0.2.10"
yfinance = "^0.2.54"
xmltodict = "^0.14.2"
google-search-results = "^2.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
nest-asyncio = "^1.6.0"


[tool.poetry.group.jupyter]
optional = true

[tool.poetry.group.jupyter.dependencies]
jupyter = "^1.0.0"


[tool.poetry.group.lint]
optional = true

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
mypy = "^1"
codespell = "^2.2.0"
black = "^24.8.0"
pylint = "^3.3.1"
types-pyyaml = "^6.0.12.20240917"
types-requests = "^2.32.0.20240914"

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,poetry.lock"

[tool.flake8]
ignore = ["E501", "W503"]

[tool.pylint]
disable = [
    "E0401",
    "C0301",
    "R0903",
    "R1710",
    "C0114",
    "R0915",
    "W1514",
    "W1203",
    "C0411",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."